+++ fawkes-crypto/src/native/mod.rs
@@ -1,4 +1,3 @@
 pub mod ecc;
 pub mod eddsaposeidon;
 pub mod poseidon;
-pub mod fast_poseidon;
